refac: x_malloc && file placement
