+++ parser/minishell.bnf
@@ -1,6 +1,3 @@
-TODO:
-- wordの定義 - 数字の認識
-
 <letter>	    ::= a|b|c|d|e|f|g|h|i|j|k|l|m|n|o|p|q|r|s|t|u|v|w|x|y|z|
                     A|B|C|D|E|F|G|H|I|J|K|L|M|N|O|P|Q|R|S|T|U|V|W|X|Y|Z
 <digit>			::= 0|1|2|3|4|5|6|7|8|9
@@ -11,75 +8,34 @@ TODO:
 
 --------------------------
 
-<redirection>	    ::= '>' <word>
-                    |	'<' <word>
-                    |	'>>' <word>
-                    |	'<<' <word>
-                    |	<number>　'>' <word> #lexer space check
-                    |	<number>　'<' <word>
-                    |	<number>　'>>' <word>
-                    |	<number>　'<<' <word>
+<command_line>	::= <pipeline> '&&' <newline> <command_line>
+				|   <pipeline> '||' <newline> <command_line>
+				|   <pipeline>
 
-<redirection_list> ::= <redirection>
-                    |  <redirection_list> <redirection>
+<pipeline>	::=  <command> '|' <newline> <pipeline>
+			|   <command>
 
-<list1> ::=   <list1> '&&' <newline_list> <list1>
-           |  <list1> '||' <newline_list> <list1>
-           |  <list1> '&' <newline_list> <list1>
-           |  <list1> ';' <newline_list> <list1>
-           |  <list1> '\n' <newline_list> <list1>
-           |  <pipeline>
-<list0> ::=   <list1> '\n' <newline_list>
-           |  <list1> '&' <newline_list>
-           |  <list1> ';' <newline_list>
-<list> ::=   <newline_list> <list0>
+<command>	::= <subshell>
+			|	<simple_command>
 
-<compound_list>     ::=  <list>
-                    |  <newline_list> <list1>
-<subshell>          ::=  '(' <compound_list> ')'
-<shell_command>     ::= <subshell>
+<subshell>	::=	'(' <compound_list> ')' <redirection_list>
+			|	'(' <compound_list> ')'
 
+<compound_list>	::=	<pipeline> '&&' <newline> <compound_list>
+           		|	<pipeline> '||' <newline> <compound_list>
+           		|	<pipeline> '\n' <newline> <compound_list>
+           		|	<pipeline>
+
+<simple_command>	::= <simple_command_element> <simple_command>
+					|   <simple_command_element>
 
 <simple_command_element>	::= <word>
 							|	<redirection>
-<simple_command>			::= <simple_command_element>
-                            |   <simple_command> <simple_command_element>
 
+<redirection_list> ::= <redirection> <redirection_list>
+                   |  <redirection>
 
-<command>   ::= <simple_command>
-            |   <shell_command>
-            |   <shell_command> <redirection_list>
-
-<newline_list> ::=  <newline_list> '\n'
-                |   (empty)
-
-<pipeline>  ::=  <pipeline> '|' <newline_list> <pipeline>
-            |   <command>
-
-<simple_list>  ::= <simple_list> '&&' <newline_list> <simple_list>
-                |   <simple_list> '||' <newline_list> <simple_list>
-                |   <pipeline>
-
---------------------------
-
-
-
---------------------------
-
-
-
-<command line>	::=	<job>
-                |	<job> '&&' <command line>
-                |	<job> '||' <command line>
-
-<job>				::=	<command> '|' <job>
-					|	<command>
-
-<command>			::=	<simple command>
-	        		|	<simple command> '<' <filename>
-	        		|	<simple command> '>' <filename>
-
-<simple command>	::=	<pathname> <token list>
-
-<token list>		::=	<token> <token list>
-					|	(EMPTY)
+<redirection>	::= <number>? '>' <word>
+				|	<number>? '<' <word>
+				|	<number>? '>>' <word>
+				|	<number>? '<<' <word>
